pyproject.toml
[tool.poetry]
name = "b2broker"
version = "0.1.0"
description = ""
authors = ["Ахметов Марат Ришатович <ahmetovmr@mts.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
djangorestframework-jsonapi = "^8.0.0"
djangorestframework = "^3.16.0"
django-filter = "^25.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
